chore: slight reduction of wasteful calls

+++ Intersect.Client.Core/Maps/ActionMessage.cs
@@ -18,7 +18,7 @@ public partial class ActionMessage : IActionMessage
 
     public int X { get; set; }
 
-    public long XOffset { get; set; }
+    public int XOffset { get; set; }
 
     public int Y { get; set; }
